+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@
 Released: ???
 
 Features:
-- 
+- Added Fiery Scuttler & Ragnarok (when fishing in hotspot), Plhlegblast (when in Plhlegblast Pool) tracking to Crimson Isle tracker.
 
 Bugfixes:
 -